# Conflicts:
#	web_server/authors/views.py
#	web_server/friendship/views.py
#	web_server/posts/models.py
#	web_server/posts/templates/posts/post.html
#	web_server/posts/views.py
#	web_server/users/models.py

+++ web_server/users/views.py
@@ -61,47 +61,36 @@ def profile(request, user_id):
     })
 
 
-def invalidate_friends(host, user_id):
-    # print("\n\n\n\n\n\n\n\n")
-    # print("from invalidate_friends")
+"""
+ this function invalidates existing friendship of give author 
+ it GETs foreign servers' https://service/<authorid: UUID>/friends/<authorid2:: URL> endpoint to retrieve latest 
+ friendship status (if given author has foreign friends)
+"""
 
+
+def invalidate_friends(host, user_id):
     author_id = url_regex.sub('', user_id)
     author_id = author_id.rstrip("/")
-    # print(author_id)
-    # print(host)
-    # print("\n\n\n\n\n\n\n\n")
     if not Friend.objects.filter(author_id=author_id).exists():
         return
     friends = Friend.objects.filter(author_id=author_id)
-    # print("\n\n\n\n\n\n\n\n")
-    # print(friends)
-    # print("\n\n\n\n\n\n\n\n")
     for friend in friends:
         splits = friend.friend_id.split("/")
         friend_host = splits[0]
         if host != friend_host:
             if Node.objects.filter(pk=friend_host).exists():
                 node = Node.objects.get(foreign_server_hostname=friend_host)
-                # quoted_author_id = quote(
-                #     author_id, safe='~()*!.\'')
                 headers = {"Content-Type": "application/json",
                            "Accept": "application/json"}
                 url = "https://{}/friends/{}".format(
                     friend.friend_id, author_id)
                 res = requests.get(url, headers=headers, auth=(
                     node.username_registered_on_foreign_server, node.password_registered_on_foreign_server))
-                # print("\n\n\n\n\n")
-                # print("this is in invalidate friends ")
-                # print(url)
-                # print(res.text, res.status_code)
-                # print("\n\n\n\n\n")
                 if res.status_code >= 200 and res.status_code < 300:
                     res = res.json()
-                    # print("\n\n\n\n\n")
-                    # print(res)
-                    # print("\n\n\n\n\n")
                     # if they are friends
-                    if not res["friends"]:
+                    pending = res.get("pending", None)
+                    if not pending and not res["friends"]:
                         if Friend.objects.filter(author_id=author_id).filter(friend_id=friend.friend_id).exists():
                             Friend.objects.filter(author_id=author_id).filter(
                                 friend_id=friend.friend_id).delete()
@@ -136,13 +125,22 @@ def view_post(request, post_path):
         return HttpResponse(f"No foreign server with hostname {host} is registered on our server.", status=404)
 
     req = node.make_api_get_request(f'posts/{post_id}')
+    # Attempt to extract the post
+    try:
+        post = req.json()['posts'][0]
+    except Exception as e:
+        print(req.content)
+        return HttpResponse(f"The foreign server returned a response, but we could not extract the post. Error: {e}",
+                            status=500)
+    # Attempt to render the post
     try:
         return render(request, 'posts/foreign_post.html', {
-            'post': req.json()['posts'][0]
+            'post': post
         })
-    except:
-        return HttpResponse("The foreign server returned a response, but it was not compliant with the specification. "
-                            "We are unable to show the post at this time", status=500)
+    except Exception as e:
+        print(post)
+        return HttpResponse(f"The post we extracted from the foreign server has missing or incorrect keys: {e}",
+                            status=500)
 
 @login_required
 def view_post_comment(request, post_path):
